Country class implemented.
The Country class that has the responsiblity of storing any country's
trading statistics that can be accessed and set via properties or when
the object was originally constructed. The class implements IComparable
and can compared against other countries by their name. ToString has
been overriden to summerise the trade statistics of each instance.