CountryParser implemented. Tests included.
CountryParser is a utility class that provides the ability to parse
countries from the supplied data file. The countries are returned from
the static function within an array. In addition to this tests have been
created with the use of the NUnit framework. The tests are quite simple
currently only testing if CountryParser functions as intented and the
comparisons made by Country IComparable is correct. These tests will
expand as development goes on and will ensure that code written is
working.