CountryBSTree started.
Country specific BSTree is a class which provides BSTree functionality
specific to the Country class. It cleans up Create and Edit requiring
only the Country class rather than both key and value. Also providing
functions such as GetTradeWith which provides all list of name of
countries that trade with the supplied country name. CountryBSTree will
continue to be improved along with comments and tests later.